+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Attributes#inspect is now filtered based on filtered_attributes,
   which defaults to `Rails.application.config.filter_parameters` in
   Rails apps.
+* #143 Changed Attributes to allocate less objects (Chris Grigg)
 
 # ActiveAttr 0.13.1 (April 25, 2019) #