+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   which defaults to `Rails.application.config.filter_parameters` in
   Rails apps.
 * #143 Changed Attributes to allocate less objects (Chris Grigg)
+* #153 Changed HaveAttributeMatcher#with_default_value_of to support
+  Procs (Yoshiyuki Hirano)
 
 # ActiveAttr 0.13.1 (April 25, 2019) #